OOPS concept and Exception Handling
Object-Oriented Programming (OOP) is a popular programming paradigm that helps organize and structure code by modeling real-world entities as objects.

An exception in Python is an event or error condition that disrupts the normal flow of a program's execution. When an exceptional situation occurs during the execution of a
Python program, Python raises an exception to signal that something unexpected or erroneous has happened. Exceptions provide a mechanism for handling errors and
responding to them in a structured way, rather than allowing the program to crash